Moved the selected works map into the fetch data request component. Fetching and displaying data is working
